requirements.txt
discord.py
requests
qrcode
pillow
python-dotenv
